==> 0.obtain_data.ipynb <==
# !pip install kaggle

import kaggle
import os, zipfile

!kaggle competitions download -c airbnb-recruiting-new-user-bookings
# after this execution, move zip file into the data folder before executing next line

pwd = os.getcwd()
dir_name = pwd+'\data\\'

os.listdir(dir_name)

for item in os.listdir(dir_name): # loop through items in dir
    if item.endswith('.zip'): # check for ".zip" extension
        file_name = dir_name+item # get full path of files
        zip_ref = zipfile.ZipFile(file_name) # create zipfile object
        zip_ref.extractall(dir_name) # extract file to dir
        zip_ref.close() # close file
        os.remove(file_name) # delete zipped file

for item in os.listdir(dir_name): # loop through items in dir
    if item.endswith('.zip'): # check for ".zip" extension
        file_name = dir_name+item # get full path of files
        zip_ref = zipfile.ZipFile(file_name) # create zipfile object
        zip_ref.extractall(dir_name) # extract file to dir
        zip_ref.close() # close file
        os.remove(file_name) # delete zipped file
