==> 2.feature_select.ipynb <==
# Feature Selection

In this section, we will focus on the session information, in which action and details are in form of many categorical variables. It is difficult to visual them one by one, hence we will apply feature reduction techniques such as LassoCV and RFE

import numpy as np
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_selection import RFE, RFECV
from sklearn.linear_model import LassoCV, LogisticRegression

session = pd.read_csv('./data/id_book.csv', index_col='id')
session.head()

## Last Action

Assume last row of each user session to be the last action of the session.

last_action = session.groupby(['user_id']).tail(1)

last_action.head()

df_fs = last_action.drop(['user_id', 'secs_elapsed'], axis=1)

df_fs_en = pd.get_dummies(df_fs, drop_first=True, prefix=df_fs.columns[:-1])

df_fs_en.shape

X = df_fs_en.drop(['book'], axis=1)
y = df_fs_en['book']

lcv = LassoCV()
lcv.fit(X, y)

lcv_mask = lcv.coef_ != 0

var = sum(lcv_mask)

rfe_lr = RFE(estimator=LogisticRegression(multi_class='multinomial',
                                          solver='lbfgs', max_iter=1000), n_features_to_select=var, step=5)
rfe_lr.fit(X, y)

rfe_rf = RFE(estimator=RandomForestClassifier(),
             n_features_to_select=var, step=10)
rfe_rf.fit(X, y)

lr_mask = rfe_lr.support_
rf_mask = rfe_rf.support_

votes = np.sum([lcv_mask, lr_mask, rf_mask], axis=0)
mask = votes == 3

X_select = X.loc[:, mask]

X_select.shape

# df = pd.merge(X_select, y, left_index=True, right_index=True)

X_select.to_csv('./data/session_df.csv')

### [Return to README](https://github.com/sittingman/airbnb_booking/blob/master/README.md)
