==> 3.data_prep.ipynb <==
# Data Preparation

The workbook performs needed steps to transform raw data into the final dataset format to be ready for machine learning purposes.

### Preparing train set

import datetime as dt

import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
import numpy as np
import pandas as pd
import seaborn as sns

plt.style.use('ggplot')
%matplotlib inline

gender = pd.read_csv('./data/age_gender_bkts.csv')
train = pd.read_csv('./data/train_users_2.csv')
test = pd.read_csv('./data/test_users.csv')
country = pd.read_csv('./data/countries.csv')
session = pd.read_csv('./data/sessions.csv')
age_gender = pd.read_csv('./data/age_gender_bkts.csv')

age_cat = age_gender['age_bucket'].unique().tolist()
age_sort = age_cat[::-1]
age_sort.insert(0, 'miss')

train['age_fill'] = train['age'].apply(
    lambda x: -1 if x > 1000 else x).fillna(-1)
train['age_gp'] = pd.cut(
    train['age_fill'], bins=np.arange(-5, 106, 5), labels=age_sort)

train['first_affiliate_tracked'].fillna('miss', inplace=True)

lang = {'en': 'eng',
        'zh': 'asia',
        'ko': 'asia',
        'ja': 'asia',
        'fr': 'eu',
        'es': 'eu',
        'de': 'eu',
        'ru': 'eu',
        'pt': 'eu'}

train['lang'] = train['language'].map(lang).fillna('other')

browser = {'Chrome': 'chrome',
           'Chrome Mobile': 'chrome',
           'Safari': 'safari',
           'Mobile Safari': 'safari',
           'Firefox': 'firefox',
           'Mobile Firefox': 'firefox',
           'IE': 'ie',
           'IE Mobile': 'ie',
           '-unknown-': 'unknown'}

train['browser'] = train['first_browser'].map(browser).fillna('other')

train['gender_stated'] = train['gender'].apply(
    lambda x: 'no' if x == '-unknown-' else 'yes')

flow_cat = ['0', '1-10', '11-20', '20+']
train['flow_cat'] = pd.cut(train['signup_flow'], bins=[
                           0, 1, 10, 20, 100], labels=flow_cat, include_lowest=True)

train.head()

col_to_keep = ['id', 'signup_method',
       'affiliate_channel', 
       'signup_app', 'age_gp', 'lang',
       'browser', 'gender_stated', 'flow_cat', 'country_destination']

df_train = train[col_to_keep]

df_train.set_index('id', inplace=True)

df_train = pd.get_dummies(df_train, prefix=df_train.columns[:-1], columns=df_train.columns[:-1], drop_first=True)

df_train.head()

### Preparing Test Set

test['age_fill'] = test['age'].apply(
    lambda x: -1 if x > 1000 else x).fillna(-1)
test['age_gp'] = pd.cut(
    test['age_fill'], bins=np.arange(-5, 106, 5), labels=age_sort)

test['first_affiliate_tracked'].fillna('miss', inplace=True)

test['lang'] = test['language'].map(lang).fillna('other')

test['browser'] = test['first_browser'].map(browser).fillna('other')

test['gender_stated'] = test['gender'].apply(
    lambda x: 'no' if x == '-unknown-' else 'yes')

test['flow_cat'] = pd.cut(test['signup_flow'], bins=[
                           0, 1, 10, 20, 100], labels=flow_cat, include_lowest=True)

df_test = test[col_to_keep[:-1]]
df_test.set_index('id', inplace=True)
df_test = pd.get_dummies(df_test, prefix=df_test.columns, columns=df_test.columns, drop_first=False)
df_test = df_test[df_train.columns[1:]]

print(len(df_test.columns), len(df_train.columns))

### Combining session data

session_df = pd.read_csv('./data/session_df.csv', index_col=0)
encoded_col = session_df.columns

last_action = session.drop(['secs_elapsed'], axis=1).groupby(['user_id']).tail(1)
last_action.set_index('user_id', inplace=True)

last_act_df = pd.get_dummies(last_action, drop_first=False, prefix=last_action.columns)

session_df = last_act_df.loc[:, encoded_col]

train_fin = df_train.merge(session_df, left_index=True, right_index=True, how='left').fillna(0)
test_fin = df_test.merge(session_df, left_index=True, right_index=True, how='left').fillna(0)

print(train_fin.shape, test_fin.shape)

train_fin.to_csv('./data/train_df.csv')
test_fin.to_csv('./data/test_df.csv')
