==> 5.sumbit.ipynb <==
# Data Submission to Kaggle

import pickle
import warnings
import numpy as np
import pandas as pd
from keras.models import load_model
from sklearn.dummy import DummyClassifier
from lightgbm import LGBMClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.preprocessing import LabelBinarizer, LabelEncoder
warnings.filterwarnings(action='ignore')

train = pd.read_csv('./data/train_df.csv', index_col=0)

test = pd.read_csv('./data/test_df.csv', index_col=0)

le = LabelEncoder()
train['target'] = le.fit_transform(train['country_destination'])
lb = LabelBinarizer()
lb.fit(train['target'])

target = train['target']
feature = train.drop(['target', 'country_destination'], axis=1)

def submit_data(name, pred):
    """function to submit data to kaggle 
    in the required form

    Args:
        name: name of the output csv file
        pred: predictions generated from algorithm

    Return:
        csv file in the required kaggle format
    """
    id_test = test.index.to_series()
    ids = []
    cts = []
    for i in range(len(id_test)):
        idx = id_test[i]
        ids += [idx] * 5
        cts += le.inverse_transform(np.argsort(pred[i])[::-1])[:5].tolist()
    sub = pd.DataFrame(np.column_stack((ids, cts)), columns=['id', 'country'])
    sub.to_csv('./data/'+name, index=False)

### Base Models

dum = DummyClassifier()
dum.fit(feature, target)
dum_pred = dum.predict_proba(test)

submit_data('dum.csv', dum_pred)

rfc = RandomForestClassifier()
rfc.fit(feature, target)
rf_pred = rfc.predict_proba(test)

submit_data('rfc_base.csv', rf_pred)

et = ExtraTreesClassifier()
et.fit(feature, target)
et_pred = et.predict_proba(test)

submit_data('et_base.csv', et_pred)

lgb = LGBMClassifier()
lgb.fit(feature, target)
lgb_pred = lgb.predict_proba(test)

submit_data('lgb_base.csv', lgb_pred)

### Tunned Models

rf_tuned = pickle.load(open('./data/rf_tuned.sav', 'rb'))
rf_tuned_pred = rf_tuned.predict_proba(test)

rf_tuned

submit_data('rf_tuned.csv', rf_tuned_pred)

et_tuned = pickle.load(open('./data/et_tuned.sav', 'rb'))
et_tuned_pred = et_tuned.predict_proba(test)

et_tuned

submit_data('et_tuned.csv', et_tuned_pred)

lgb_tuned = pickle.load(open('./data/lgb_tuned.sav', 'rb'))
lgb_tuned_pred = lgb_tuned.predict_proba(test)

lgb_tuned

submit_data('lgb_tuned.csv', lgb_tuned_pred)

### Deep Learning Model

deep_mod = load_model('./data/deep_learn.tf')

deep_mod_pred = deep_mod.predict_proba(test)

submit_data('deep_learn.csv', deep_mod_pred)
